refactor(suite-native): Screen header
