+++ suite-native/module-settings/src/screens/SettingsCustomizationScreen.tsx
@@ -1,12 +1,19 @@
 import React from 'react';
 
 import { Screen, ScreenHeader } from '@suite-native/navigation';
-import { Box } from '@suite-native/atoms';
+import { Box, Text } from '@suite-native/atoms';
 
 import { ColorSchemePicker } from '../components/ColorSchemePicker';
 
 export const SettingsCustomizationScreen = () => (
-    <Screen header={<ScreenHeader title="Customization" />}>
+    <Screen
+        header={
+            <ScreenHeader
+                titleComponent={<Text variant="titleSmall">Customization</Text>}
+                hasGoBackIcon
+            />
+        }
+    >
         <Box marginHorizontal="medium">
             <ColorSchemePicker />
         </Box>
