+++ suite-native/module-settings/src/screens/SettingsLocalisationScreen.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
 
-import { VStack } from '@suite-native/atoms';
+import { VStack, Text } from '@suite-native/atoms';
 import { Screen, ScreenHeader } from '@suite-native/navigation';
 
 import { CurrencySelector } from '../components/CurrencySelector';
 
 export const SettingsLocalisationScreen = () => (
-    <Screen header={<ScreenHeader title="Localisation" />}>
+    <Screen
+        header={<ScreenHeader titleComponent={<Text variant="titleSmall">Localisation</Text>} />}
+    >
         <VStack spacing={12}>
             <CurrencySelector />
         </VStack>
