+++ suite-native/navigation/src/index.ts
@@ -5,4 +5,3 @@ export * from './config';
 export * from './components/TabBar';
 export * from './components/Screen';
 export * from './components/ScreenHeader';
-export * from './components/ScreenHeaderWithIcons';
